🚧wip (api): split api helpers into server and client files
