+++ src/pages/api/students/[id].ts
@@ -1,7 +1,7 @@
 import { ReasonPhrases, StatusCodes } from "http-status-codes";
 
 import { connect } from "db";
-import { routeWrapper } from "api";
+import { routeWrapper } from "api/server";
 import { USER_ID_COOKIE } from "utils/constants";
 import { StudentModel, TeacherStaffModel } from "db/models";
