+++ src/pages/api/teachers/[id]/students.ts
@@ -2,7 +2,7 @@ import { formatDistanceToNowStrict } from "date-fns";
 import { ReasonPhrases, StatusCodes } from "http-status-codes";
 
 import { connect } from "db";
-import { routeWrapper } from "api";
+import { routeWrapper } from "api/server";
 import { ClassModel, SubjectModel, StudentModel, TermModel } from "db/models";
 
 import type { NextApiRequest, NextApiResponse } from "next";
