+++ src/pages/api/terms/[id].ts
@@ -1,7 +1,7 @@
 import { ReasonPhrases, StatusCodes } from "http-status-codes";
 
 import { connect } from "db";
-import { NotFoundError, routeWrapper } from "api";
+import { NotFoundError, routeWrapper } from "api/server";
 import { TermModel } from "db/models";
 
 import type { NextApiRequest, NextApiResponse } from "next";
