+++ src/pages/login.tsx
@@ -5,7 +5,7 @@ import { Fragment, useState } from "react";
 import Head from "next/head";
 
 import { useToast } from "components";
-import { fetchAPIEndpoint } from "api";
+import { fetchAPIEndpoint } from "api/client";
 import { JWT_COOKIE_TOKEN, REDIRECT_QUERY } from "utils/constants";
 import { Checkbox, Input, Password, Select } from "components/Form";
