+++ src/pages/staff-teacher/classes/create.tsx
@@ -2,7 +2,7 @@ import Head from "next/head";
 import { Fragment, useState } from "react";
 
 import { useToast } from "components";
-import { fetchAPIEndpoint } from "api";
+import { fetchAPIEndpoint } from "api/client";
 import { LoadingIcon } from "components/Icons";
 import { Input, Users } from "components/Form";
